requirements.txt
annotated-types==0.7.0
anyio==4.5.0
bcrypt==4.2.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
fastapi-jwt-auth==0.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
mysql-connector-python==9.0.0
numpy==2.1.3
passlib==1.7.4
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==1.7.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.31.1
